backend/requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
graphviz==0.20.3
h11==0.14.0
idna==3.10
requests==2.32.3
urllib3==2.3.0
uvicorn==0.34.0
fastapi==0.115.12
py-mon==2.0.5
transformers==4.49.0
geopy==2.4.1
mysql-connector-python==9.2.0
torch==2.6.0
numpy==2.0.2
shapely==2.1.0
geopandas==1.0.1
pydantic==1.10.11
hf-xet==1.0.4